Add files via upload
Added a way to turn off the displays in the even that we're using them in a container that can't support GUI's. Also set the default image topic to /bebop/image_raw rather than /image_raw. Tweaked the levels a little to better find the orange tennisball, but those are subject to change.